add status tracking through sqlite3 database
As per discussion in #4, using sqlite3 as a database for storage of output seems like a good choice. The current structure is as follows:

1. When a new Pipeline is instantiated, a connection is made to the database indicated in the relevant configuration (right now that lives as the value of a "statusdb" key).
2. A new row is added to the database with that pipeline's passed name and display name.
3. If the Pipeline fails at any point, that row is updated to reflect the Exception message
4. If the Pipeline succeeds, the status is updated to be successful and some additional metadata (number of rows processed, e.g.) is added
5. When the Pipeline run stops, the connection to the database is closed.

As a part of this, you can also pass a sqlite3 connection object into the pipeline, which will leave the connection open after the run finishes, which is very useful for in-memory sqlite3 testing capabilities. I've also added a small script that lets the user create new databases by running the following command:

```bash
python create_db.py --server "name of server in config file" --drop [optionally, to drop the status table before recreating it]
```