add connector base instance
the goal here is to separate two primary concerns:

1) *connecting* to some resource
2) *reading* from that resource

in the simplist case, connecting to a resource is as simple as calling open() on the resource's path. however, for more complex cases, more advanced connection might be needed. additionally, while it may originally appear as though the two concerns are completely orthagonal, it seems as though they are a bit more tied together than we might hope. therefore, a method has been introduced on the Extractor class named `process_connection`, which allows an individual extractor class to process the connection returned by `Connector.connect`. For example, the `CSVExtractor` applies a `csv.reader` to the returned file object, allowing for the benefits of using it (such as handling quote characters etc.). this provides the additional benefit of only having to stream through the file once.

thanks to @wpietri on this commit